Fix vscode sanitize (#717)
* Sanitize vscode url

* Add missing component files
