+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Fixed sanitizing VsCode callbackUrl](https://github.com/multiversx/mx-sdk-dapp/pull/717)
 - [Added missing endpoints](https://github.com/multiversx/mx-sdk-dapp/pull/716)
-
 ## [[v2.10.15]](https://github.com/multiversx/mx-sdk-dapp/pull/715)] - 2023-03-28
 - [Added missing component for `DappProvider` instead of `index`](https://github.com/multiversx/mx-sdk-dapp/pull/715)
