+++ src/UI/SignTransactionsModals/SignWithExtensionModal/SignWithExtensionModal.tsx
@@ -1,0 +1,1 @@
+export * from './SignWithExtensionModal';
