+++ src/UI/SignTransactionsModals/SignWithExtraModal/SignWithExtraModal.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { useSignTransactions } from 'hooks/transactions/useSignTransactions';
+
+export function SignWithExtraModal() {
+  useSignTransactions();
+  return <></>;
+}
