+++ src/UI/SignTransactionsModals/SignWithExtraModal/index.tsx
@@ -1,7 +1,1 @@
-import React from 'react';
-import { useSignTransactions } from 'hooks/transactions/useSignTransactions';
-
-export function SignWithExtraModal() {
-  useSignTransactions();
-  return <></>;
-}
+export * from './SignWithExtraModal';
