+++ src/UI/SignTransactionsModals/SignWithLedgerModal/SignWithLedgerModal.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { SignModalPropsType } from 'types';
+import { SignWithDeviceModal } from '../SignWithDeviceModal';
+
+export const SignWithLedgerModal = (props: SignModalPropsType) => {
+  return (
+    <SignWithDeviceModal
+      {...props}
+      title={props.title || 'Confirm on Ledger'}
+      className={props.className || 'dapp-ledger-modal'}
+    />
+  );
+};
