+++ src/UI/SignTransactionsModals/SignWithLedgerModal/index.tsx
@@ -1,13 +1,1 @@
-import React from 'react';
-import { SignModalPropsType } from 'types';
-import { SignWithDeviceModal } from '../SignWithDeviceModal';
-
-export const SignWithLedgerModal = (props: SignModalPropsType) => {
-  return (
-    <SignWithDeviceModal
-      {...props}
-      title={props.title || 'Confirm on Ledger'}
-      className={props.className || 'dapp-ledger-modal'}
-    />
-  );
-};
+export * from './SignWithLedgerModal';
