+++ src/UI/SignTransactionsModals/SignWithWalletConnectModal/SignWithWalletConnectModal.tsx
@@ -1,0 +1,1 @@
+export * from './SignWithWalletConnectModal';
