+++ src/utils/sanitizeCallbackUrl.ts
@@ -3,6 +3,7 @@ export const sanitizeCallbackUrl = (
   vulnerableItems: string[] = ['address']
 ) => {
   const url = new URL(targetURL);
+
   const params = new URLSearchParams(url.search);
 
   vulnerableItems.forEach((vulnerableItem) => params.delete(vulnerableItem));
@@ -11,6 +12,10 @@ export const sanitizeCallbackUrl = (
 
   const pathname =
     url.pathname === '/' && !targetURL.endsWith('/') ? '' : url.pathname;
+
+  if (url.protocol === 'vscode:') {
+    return targetURL;
+  }
 
   return `${url.origin}${pathname}${questionMark}${params.toString()}`;
 };
